Правки обучения Добавил небольшой скриптик для анализа потребления памяти и времени на каждом из слоев. Перевел вычисление вейвлетов на гпу и ускорил тем самым в 4-5 раз

+++ neurostreak/dataset/dataset.py
@@ -39,4 +39,4 @@ class NeuroStreakDataset(Dataset):
 
         signal = img[row]
 
-        return torch.tensor(signal), torch.tensor(template) , torch.tensor(labels).to(dtype=torch.float32)
+        return torch.tensor(signal).to(dtype=torch.float32), torch.tensor(template).to(dtype=torch.float32) , torch.tensor(labels).to(dtype=torch.float32)
